+++ maxillo/views/classification.py
@@ -3,7 +3,6 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required, user_passes_test
 from django.contrib import messages
 from django.http import JsonResponse, HttpResponse
-from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_POST
 import json
 import os
@@ -17,7 +16,6 @@ logger = logging.getLogger(__name__)
 
 @login_required
 @require_POST
-@csrf_exempt
 def update_classification(request, patient_id):
     """AJAX endpoint for instant classification updates"""
     domain_models = get_domain_models(request)